SHOWDATA externe
