+++ TD_React/package.json
@@ -16,7 +16,8 @@
     "@mui/material": "^6.3.0",
     "file-saver": "^2.0.5",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router": "^7.1.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
